+++ jwst/saturation/saturation.py
@@ -54,6 +54,8 @@ def flag_saturation(input_model, ref_model, n_pix_grow_sat):
     gdq = output_model.groupdq
     pdq = output_model.pixeldq
 
+    zframe = input_model.zeroframe if input_model.meta.exposure.zero_frame else None
+
     # Extract subarray from saturation reference file, if necessary
     if reffile_utils.ref_matches_sci(input_model, ref_model):
         sat_thresh = ref_model.data
@@ -65,8 +67,9 @@ def flag_saturation(input_model, ref_model, n_pix_grow_sat):
         sat_dq = ref_sub_model.dq.copy()
         ref_sub_model.close()
 
-    gdq_new, pdq_new = flag_saturated_pixels(data, gdq, pdq, sat_thresh,
-                                             sat_dq, ATOD_LIMIT, dqflags.pixel, n_pix_grow_sat)
+    gdq_new, pdq_new, _ = flag_saturated_pixels(
+        data, gdq, pdq, sat_thresh, sat_dq, ATOD_LIMIT, dqflags.pixel,
+        n_pix_grow_sat=n_pix_grow_sat, zframe=zframe)
 
     # Save the flags in the output GROUPDQ array
     output_model.groupdq = gdq_new
@@ -140,6 +143,13 @@ def irs2_flag_saturation(input_model, ref_model, n_pix_grow_sat):
 
     flagarray = np.zeros(data.shape[-2:], dtype=groupdq.dtype)
     flaglowarray = np.zeros(data.shape[-2:], dtype=groupdq.dtype)
+
+    if input_model.meta.exposure.zero_frame:
+        zflagarray = np.zeros(data.shape[-2:], dtype=groupdq.dtype)
+        zflaglowarray = np.zeros(data.shape[-2:], dtype=groupdq.dtype)
+
+    # import ipdb; ipdb.set_trace()
+
     for ints in range(nints):
         for group in range(ngroups):
             # Update the 4D groupdq array with the saturation flag.
@@ -169,6 +179,25 @@ def irs2_flag_saturation(input_model, ref_model, n_pix_grow_sat):
             # for A/D floor, the flag is only set of the current plane
             np.bitwise_or(groupdq[ints, group, :, :], flaglowarray,
                           groupdq[ints, group, :, :])
+
+        # Process ZEROFRAME.  Instead of setting a ZEROFRAME DQ array, data
+        # in the ZEROFRAME that is flagged will be set to 0.
+        if input_model.meta.exposure.zero_frame:
+            zplane = input_model.zeroframe[ints, :, :].copy()
+            zdq = np.zeros(groupdq.shape[-2:], dtype=groupdq.dtype)
+            ztemp = x_irs2.from_irs2(zplane, irs2_mask, detector)
+
+            zflag_temp = np.where(ztemp >= sat_thresh, SATURATED, 0)
+            zflaglow_temp = np.where(ztemp <= 0, AD_FLOOR | DONOTUSE, 0)
+            x_irs2.to_irs2(zflagarray, zflag_temp, irs2_mask, detector)
+            x_irs2.to_irs2(zflaglowarray, zflaglow_temp, irs2_mask, detector)
+
+            np.bitwise_or(zdq[:, :], zflagarray, zdq[:, :])
+            np.bitwise_or(zdq[:, :], zflaglowarray, zdq[:, :])
+
+            zplane[zdq != 0] = 0.
+            output_model.zeroframe[ints, :, :] = zplane[:, :]
+            del zdq
 
     # Save the flags in the output GROUPDQ array
     output_model.groupdq = groupdq